Split setup post-apply verifier models from logic
