+++ Docs/reviewer/onboarding-unification-tracker.md
@@ -43,6 +43,7 @@ This file tracks onboarding/cleanup/maintenance unification across CLI, Web, and
 | Web autodetect fail-fast enum wire guard | Completed | IX CLI Web + IX Tests | Unknown `SetupOnboardingCheckStatus` values now throw during web payload projection and regression tests assert fail-fast behavior. |
 | Setup test maintainability split (autodetect/post-apply groups) | Completed | IX Tests | Split large `Program.Setup.cs` into focused partial files (`Program.Setup.AutodetectContract.cs`, `Program.Setup.PostApply.cs`) to reduce monolith size and keep static-analysis maintainability checks actionable. |
 | Post-apply verification constants for check names/statuses | Completed | IX CLI + IX Tests | Added shared `SetupPostApplyCheckNames`/`SetupPostApplyCheckValues` constants and switched post-apply tests to use them; also replaced parse-based fixed timestamp with deterministic constructor-based `DateTimeOffset`. |
+| Post-apply verifier maintainability split | Completed | IX CLI | Split `SetupPostApplyVerify.cs` models/constants into `SetupPostApplyVerify.Models.cs` so verifier logic stays under maintainability thresholds while preserving behavior. |
 | Docs: canonical path matrix + Bot contract-check diagram | Completed | Docs | `Docs/reviewer/web-onboarding.md` now includes per-path auth/operation matrix and Mermaid flow for tool-driven parity verification. |
 | Docs: diagrams and flow updates | Completed | Docs | Added path-first flow docs and Mermaid diagrams in onboarding pages. |
 | Website FAQ/data updates | Completed | Website | Updated FAQ/features/how-it-works for path-first + auto-detect positioning. |
